jointlly_backend/requirements.txt
# FastAPI and ASGI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0  # PostgreSQL async driver (optional if using MySQL)
aiomysql==0.2.0  # MySQL async driver
pymysql==1.1.1   # MySQL sync driver (for Alembic)
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Validation
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# Payments
razorpay==1.4.2

# Utilities
python-dotenv==1.0.1
httpx==0.27.2

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
